+++ src/commands/Owner/sync.ts
@@ -7,7 +7,7 @@ import { exec } from 'node:child_process';
 
 import { embedBuilder } from '#utils/classes/embeds';
 import { prompter } from '#utils/classes/prompter';
-import { MiscEmotes } from '#utils/emotes';
+import { MiscEmotes } from '#utils/constants';
 
 @ApplyOptions<Command.Options>({
 	description: 'Pulls code changes from github and reloads',
