--- src/lib/utils/classes/formatter.ts
@@ -1,5 +1,0 @@
-import { blueBright, bold } from 'colorette';
-
-export function formatShardHeader(shardId: number, suffix: string) {
-	return `${bold(blueBright(`[SHARD ${shardId}]`))} ${suffix}`;
-}
