--- src/lib/utils/emotes.ts
@@ -1,4 +1,0 @@
-export enum MiscEmotes {
-	Success = '<:success:959344194785271809>',
-	Error = '<:error:959359264533643274>'
-}
