+++ src/lib/utils/util.ts
@@ -2,11 +2,6 @@ import type { Interaction } from 'discord.js';
 import { DiscordAPIError, Message } from 'discord.js';
 import type { RESTJSONErrorCodes } from 'discord-api-types/v10';
 
-export function centerAlign(value: string, width: number) {
-	const padding = Math.floor((width - value.length) / 2);
-	return `${' '.repeat(padding)}${value}${' '.repeat(padding)}`;
-}
-
 export function isMessage(messageOrInteraction: Message | Interaction): messageOrInteraction is Message {
 	return messageOrInteraction instanceof Message;
 }
