+++ src/listeners/shards/shard-disconnect.ts
@@ -3,7 +3,7 @@ import { Events, Listener } from '@sapphire/framework';
 import { red } from 'colorette';
 import type { CloseEvent } from 'discord.js';
 
-import { formatShardHeader } from '#utils/classes/formatter';
+import { formatShardHeader } from '#root/lib/utils/formatter';
 
 @ApplyOptions<Listener.Options>({
 	name: 'ShardDisconnect',
